pulling

+++ chess_agents.py
@@ -149,7 +149,7 @@ class MinimaxAgent(Agent):
             return v
 
 class AlphaBetaAgent(Agent):
-    def get_move(self, game_state, return_value=False):
+    def get_move(self, game_state, return_value=False, parallelize=False):
         """
         Return minimax move using self.depth, self.eval_func, and alpha-beta pruning.
         """
@@ -157,15 +157,21 @@ class AlphaBetaAgent(Agent):
         if len(moves) == 0:
             return None
 
-        get_ab_value = partial( self._alpha_beta_value, board=game_state.board,
-                                alpha=-99999, beta=99999, depth=0,
-                                color=self.color)
+        if parallelize:
+            get_ab_value = partial( self._alpha_beta_value, board=game_state.board,
+                                    alpha=-99999, beta=99999, depth=0,
+                                    color=self.color)
 
-        p = Pool(8)
-        values = p.map_async(get_ab_value, moves).get(99999)
-        p.terminate()
+            p = Pool(8)
+            values = p.map_async(get_ab_value, moves).get(99999)
+            p.terminate()
 
-        values = {mv: v for mv, v in zip(moves, values)}
+            values = {mv: v for mv, v in zip(moves, values)}
+
+        else:
+            values = {}
+            for move in moves:
+                values[move] = self.get_value(move, game_state.board, 0, self.color)
 
         # return action with max utility,
         # random action if there's a tie
